spm: keep it building for nRF9160-DK
The devicetree overlay is enabling uart2 without setting its
current-speed property, which causes an error in the driver code,
which is attempting to instantiate (and initialize) a device and
therefore needs to know its baud rate.

Fix that by switching to using the HAL Kconfig option to allow access
to the CMSIS NRF_UARTE2 macro. That's all the spm really needs, as
it's not using the Zephyr device model at all here.

This same change should be investigated for the rest of the devicetree
overlay for this board, as the PWM definitions similarly do not seem
necessary, and removing them could same some boot time and leave the
peripherals untouched since Zephyr driver init routines for them won't
be run.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>