tree-wide: adjust CONFIG_ENTROPY_NAME usages
This is gone now; the chosen zephyr,entropy device is the replacement.
Get this label from the common test code.

In the case of entropy_ll_nrfxlib.c, I'm assuming this device
instantiation was because the driver is intended to be the entropy
device exactly when it's compiled in, so just name the device
according to the chosen node. Hopefully that is OK.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>