lib/edge_impulse/CMake: Ensure EI_URI is string
This patch made sure that EI_URI is threaten as string by CMake.
zephyr-rtos changed so CONFIG_EDGE_IMPULSE_URI is transferred
with semicolon - so it is recognized as a list of string instead of
a string like it was before.

This patch adds these strings concatenation, so EI_URI stays the string
even that.

Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>