bl_storage: fw_info/counters port to RRAMC
Changed the implementation of bl_storage to be compatible with RRAMC and
FLASH/NVMC. Due to the limitations in the RRAM controller OTP can only
be written to by words not half-words as it can be in NVMC.

This means we need to change the way we handle certain cases in
`bl_storage.c`, `bl_validation.c` and `fw_info.c`. This commit moves
some of the abstractions into two separate files `bl_storage_nvmc.c` and
`bl_storage_rramc.c`. However there are multiple other changes needed
which are done by using `#ifdef`'s.

Ref. NCSDK-25306

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>