cmake: provision: Change provision_hex to account for larger LCS
Changed the provisioning script to account for larger Life Cycle State
structures. Since we can only do word writes in nRF54LX life cycle
states needs to be stored in words. This increased the LCS struct by 1
word requiring us to change the way the provisioning script works.

This change adds the LCS size as a value to the CMake file which uses
the SOC to decide which size the LCS struct is which is then passed to
the python script through the `--lcs-state-size` argument.

Ref. NCSDK-25306

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>